\name{EMPIRmed.regress}
\alias{EMPIRmed.regress}
\title{ Derivatives of the Grid of the Bivariate Emprical Copula }
\description{
Generate median regression from the gridded inversion of the bivariate empirical copula.
}
\usage{
EMPIRmed.regress(...)
}
\arguments{
  \item{...}{Arguments to pass to \code{\link{EMPIRqua.regress2}}.}
}
\value{
  The gridded values of
}
\author{ W.H. Asquith }
\seealso{ \code{\link{EMPIRgridderinv}}, \code{\link{EMPIRqua.regress}}, \code{\link{EMPIRmed.regress2}}, \code{\link{EMPIRmed.regress2}}
}
\examples{
# See examples under EMPIRqua.regress
}
\keyword{distribution}
\keyword{multivariate}
