\encoding{utf8}
\name{EMPIRmed.regress}
\alias{EMPIRmed.regress}
\title{ Median Regression of the Grid of the Bivariate Empirical Copula for V with respect to U }
\description{
Perform \emph{median regression} from the gridded inversion of the \emph{bivariate empirical copula} of \eqn{V} with respect to \eqn{U}.
}
\usage{
EMPIRmed.regress(...)
}
\arguments{
  \item{...}{Arguments to pass to \code{\link{EMPIRqua.regress}}.}
}
\value{
  The gridded values of the median regression of \eqn{V} with respect to \eqn{U}.
}
\author{ W.H. Asquith }
\seealso{ \code{\link{EMPIRgridderinv}}, \code{\link{EMPIRqua.regress}}, \code{\link{EMPIRmed.regress2}}, \code{\link{EMPIRmed.regress2}}
}
\examples{
# See examples under EMPIRqua.regress
}
\keyword{empirical copula}
\keyword{empirical copula (median regression)}
\keyword{empirical copula (quantile regression)}
