\name{EMPIRmed.regress2}
\alias{EMPIRmed.regress2}
\title{ Median Regression of the Grid of the Bivariate Empirical Copula for U with respect to V}
\description{
Perform \emph{median regression} from the gridded inversion of the \emph{bivariate empirical copula} of \eqn{U} with respect to \eqn{V}.
}
\usage{
EMPIRmed.regress2(...)
}
\arguments{
  \item{...}{Arguments to pass to \code{\link{EMPIRqua.regress2}}.}
}
\value{
  The gridded values of the median regression of \eqn{U} with respect to \eqn{V}.
}
\author{ W.H. Asquith }
\seealso{ \code{\link{EMPIRgridderinv2}}, \code{\link{EMPIRqua.regress}}, \code{\link{EMPIRmed.regress}}, \code{\link{EMPIRmed.regress2}}
}
\examples{
# See examples under EMPIRqua.regress
}
\keyword{empirical copula}
\keyword{empirical copula (median regression)}
\keyword{empirical copula (quantile regression)}
